Make a variants report as named list of tables

Make a variants report as named list of tables

Usage

make_report(vars, subs, fus = DataFrame(), tiled)

Arguments

vars

Variant results from `annotate_coding()`

subs

Variants within mutation context from `subset_context()`

fus

Variants within fusion context from `subset_context_fusions()`

...

Parameters passed to `pep_tile` (eg. tile_size, tile_ov)

ctx_codons

Number of codonds for sequence context